[feat] 스레드 추가 로직 service로 분리 및 리턴값에 insertId 리턴
